+++ src/main/java/io/crate/jmx/recorder/QueryStats.java
@@ -35,6 +35,7 @@ public class QueryStats implements Recorder {
     private static final Pattern QUERIES_PER_SECONDS_PATTERN = Pattern.compile("(.+)QueryFrequency");
     private static final Pattern QUERIES_DURATION_PATTERN = Pattern.compile("(.+)QueryAverageDuration");
     private static final Pattern QUERIES_TOTAL_COUNT_PATTERN = Pattern.compile("(.+)QueryTotalCount");
+    private static final Pattern QUERIES_AFFECTED_ROW_COUNT_PATTERN = Pattern.compile("(.+)QueryAffectedRowCount");
     private static final Pattern QUERIES_FAILED_COUNT_PATTERN = Pattern.compile("(.+)QueryFailedCount");
     private static final Pattern QUERIES_SUM_OF_DURATIONS_PATTERN = Pattern.compile("(.+)QuerySumOfDurations");
 
@@ -74,6 +75,17 @@ public class QueryStats implements Recorder {
                     beanValue,
                     "The total number of queries that were executed for a given query type.",
                     metricSampleConsumer);
+            return true;
+        }
+        matcher = QUERIES_AFFECTED_ROW_COUNT_PATTERN.matcher(attrName);
+        if (matcher.matches()) {
+            recordBean(
+                domain,
+                "query_affected_row_count",
+                matcher.group(1),
+                beanValue,
+                "The total number of affected rows of all statement executions for a given query type.",
+                metricSampleConsumer);
             return true;
         }
         matcher = QUERIES_FAILED_COUNT_PATTERN.matcher(attrName);